Fix inconsistent training results with RGBA/PNG images
The training relies on PIL to resize the input images and extracts the resized alpha
to mask the rendered image during training. Since PIL pre-multiplies the resized RGB
with the resized alpha, the training produces different Gaussian points depending on
whether the input get resized or not. Moreover, the extracted alpha channel from PIL
is not perfectly binarized, causing floaters around the edges.